requirements.txt
pyspark==3.5.0
boto3==1.28.63
pyyaml==5.4.1